First stab at reducing startup time
