+++ src/pip/_internal/commands/check.py
@@ -1,9 +1,6 @@
 import logging
 
 from pip._internal.cli.base_command import Command
-from pip._internal.operations.check import (
-    check_package_set, create_package_set_from_installed,
-)
 
 logger = logging.getLogger(__name__)
 
@@ -16,6 +13,9 @@ class CheckCommand(Command):
     summary = 'Verify installed packages have compatible dependencies.'
 
     def run(self, options, args):
+        from pip._internal.operations.check import (
+            check_package_set, create_package_set_from_installed,
+        )
         package_set, parsing_probs = create_package_set_from_installed()
         missing, conflicting = check_package_set(package_set)
