+++ src/pip/_internal/commands/search.py
@@ -2,22 +2,12 @@ from __future__ import absolute_import
 
 import logging
 import sys
-import textwrap
-from collections import OrderedDict
-
-from pip._vendor import pkg_resources
-from pip._vendor.packaging.version import parse as parse_version
-# NOTE: XMLRPC Client is not annotated in typeshed as on 2017-07-17, which is
-#       why we ignore the type on this import
-from pip._vendor.six.moves import xmlrpc_client  # type: ignore
 
 from pip._internal.cli.base_command import Command
-from pip._internal.cli.status_codes import NO_MATCHES_FOUND, SUCCESS
-from pip._internal.download import PipXmlrpcTransport
-from pip._internal.exceptions import CommandError
 from pip._internal.models.index import PyPI
-from pip._internal.utils.compat import get_terminal_size
-from pip._internal.utils.logging import indent_log
+
+# NOTE: XMLRPC Client is not annotated in typeshed as on 2017-07-17, which is
+#       why we ignore the type on this import
 
 logger = logging.getLogger(__name__)
 
@@ -42,6 +32,12 @@ class SearchCommand(Command):
         self.parser.insert_option_group(0, self.cmd_opts)
 
     def run(self, options, args):
+        # NOTE: XMLRPC Client is not annotated in typeshed as on 2017-07-17, which is
+        #       why we ignore the type on this import
+        from pip._internal.cli.status_codes import NO_MATCHES_FOUND, SUCCESS
+        from pip._internal.exceptions import CommandError
+        from pip._internal.utils.compat import get_terminal_size
+
         if not args:
             raise CommandError('Missing required argument (search query).')
         query = args
@@ -58,6 +54,8 @@ class SearchCommand(Command):
         return NO_MATCHES_FOUND
 
     def search(self, query, options):
+        from pip._internal.download import PipXmlrpcTransport
+        from pip._vendor.six.moves import xmlrpc_client  # type: ignore
         index_url = options.index
         with self._build_session(options) as session:
             transport = PipXmlrpcTransport(index_url, session)
@@ -72,6 +70,7 @@ def transform_hits(hits):
     packages with the list of versions stored inline. This converts the
     list from pypi into one we can use.
     """
+    from collections import OrderedDict
     packages = OrderedDict()
     for hit in hits:
         name = hit['name']
@@ -97,6 +96,13 @@ def transform_hits(hits):
 def print_results(hits, name_column_width=None, terminal_width=None):
     if not hits:
         return
+
+    import textwrap
+    from pip._vendor import pkg_resources
+    # NOTE: XMLRPC Client is not annotated in typeshed as on 2017-07-17, which is
+    #       why we ignore the type on this import
+    from pip._internal.utils.logging import indent_log
+
     if name_column_width is None:
         name_column_width = max([
             len(hit['name']) + len(highest_version(hit.get('versions', ['-'])))
@@ -132,4 +138,7 @@ def print_results(hits, name_column_width=None, terminal_width=None):
 
 
 def highest_version(versions):
+    from pip._vendor.packaging.version import parse as parse_version
+    # NOTE: XMLRPC Client is not annotated in typeshed as on 2017-07-17, which is
+    #       why we ignore the type on this import
     return max(versions, key=parse_version)
