+++ tests/unit/test_utils.py
@@ -27,10 +27,11 @@ from pip._internal.utils.glibc import check_glibc_version
 from pip._internal.utils.hashes import Hashes, MissingHashes
 from pip._internal.utils.misc import (
     call_subprocess, egg_link_path, ensure_dir, format_command_args,
-    get_installed_distributions, get_prog, normalize_path, redact_netloc,
+    get_installed_distributions, normalize_path, redact_netloc,
     redact_password_from_url, remove_auth_from_url, rmtree,
     split_auth_from_netloc, untar_file, unzip_file,
 )
+from pip._internal.utils.misc_fast_import import get_prog
 from pip._internal.utils.packaging import check_dist_requires_python
 from pip._internal.utils.temp_dir import AdjacentTempDirectory, TempDirectory
 from pip._internal.utils.ui import SpinnerInterface
